Dotween added, and delete system fixed

+++ Assets/Scripts/Components/BodyArmor.cs
@@ -2,6 +2,6 @@
 {
     public struct BodyArmor
     {
-        public float Armor;
+        public int Armor;
     }
 }
